+++ src/chanjo2/models/pydantic_models.py
@@ -37,6 +37,11 @@ class IntervalType(str, Enum):
     EXONS = "exons"
     CUSTOM = "custom_intervals"
 
+    @staticmethod
+    def get_enum_values() -> List[str]:
+        """Returns the values of the available interval types."""
+        return [member.value for member in IntervalType]
+
 
 class TranscriptTag(str, Enum):
     REFSEQ_MRNA = "refseq_mrna"